// apps/rma-app/src/app/modules/transaction/enum/filter-value.enum.ts
export enum CategoryValue {
    All_CATEGORY = 'ALL',
    MCRC = 'MCRC',
    DEFENSIVE = 1,
    CONSERVATIVE = 2,
    BALANCED = 3,
    GROWTH = 4,
    AGGRESSIVE = 5
}


export enum ClassValue {
    ALL_CLASSES = 'All',
    ALTERNATIVE = 'ALTERNATIVE',
    LOCAL_EQUITY = 'LOCAL EQUITY',
    FIXED_INCOME = 'FIXED INCOME',
    REGIONAL_EQUITY = 'REGIONAL EQUITY',
    GLOBAL_EQUITY = 'GLOBAL EQUITY',
    CASH = 'CASH'
}

export enum FundTypeValue {
    ALL_FUND = 'All',
    CONVENTIONAL = 'C',
    SHARIAH_COMPLAINT = 'I',
    CIMB_FOCUS_FUNDS = 'FF'
}